LayoutTests/platform/ios-simulator-12-wk2/imported/w3c/web-platform-tests/preload/link-header-preload-imagesrcset-expected.txt

CONSOLE MESSAGE: The resource http://localhost:8800/preload/resources/square.png?from-header&2x was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
CONSOLE MESSAGE: The resource http://localhost:8800/preload/resources/square.png?from-header&800 was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
CONSOLE MESSAGE: The resource http://localhost:8800/preload/resources/square.png?from-header&600 was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
CONSOLE MESSAGE: line 2659: Error: assert_equals: resources/square.png?from-header&1x expected 1 but got 0

FAIL Makes sure that Link headers preload images with imagesrcset/imagesizes attributes. Error: assert_equals: resources/square.png?from-header&1x expected 1 but got 0
